Expose Service as NodePort
When exposing harbor using an Ingress, the Services are created as
type `ClusterIP` by default.

Ingresses like F5 are external to the cluster, so these Services that
are used by the Ingress must be exposed as type `NodePort`.

Add helm values to set this for `core` and `portal` service, and default
these values to `ClusterIP`.

Closes #1074